Add comments to CoqFile. Turn AuxFile to private class
