+++ pylspclient/lsp_client.py
@@ -1,6 +1,7 @@
 from pylspclient import lsp_structs
 from pylspclient.lsp_endpoint import LspEndpoint
 
+
 class LspClient(object):
     def __init__(self, lsp_endpoint: LspEndpoint):
         """
